Refactor DropTable: moved to statements.py and added DropTableParser
